Fix issues caused by differences in how Stripe and our system represent currency
Stripe uses integers and 1 is equivalent to $.01 (1 cent).
In some of our database tables (the Transaction tables), we follow Stripe's
interpretation. However, in others (Company and Project tables), funds are
represented as floats where 0.01 is equivalent to $.10.

These changes should fix calculations that did not reflect the differences
described above.